move common aadhaar functions to msdk
